backend/function_trigger/requirements.txt
# Azure Function trigger dependencies (separate from FastAPI container).
# The Function runtime uses these; the FastAPI container does not.
azure-functions>=1.25.0
azure-cosmos>=4.6.0
azure-servicebus>=7.12.0
azure-identity>=1.16.0
pydantic>=2.7.1
pydantic-settings>=2.2.0
structlog>=24.1.0
